requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic-settings==2.8.1
aiogram==3.19.0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
cryptography==45.0.2
sqlmodel==0.0.21
feedparser==6.0.11
PyMuPDF==1.24.5
openpyxl==3.1.2
apscheduler==3.10.4
anthropic==0.40.0